chore: bundle pre-existing work-tree changes (credit, import, supabase, PO, i18n)
Uncommitted changes present in the working tree that were NOT authored or
verified in this session, committed as-is at the user's request:
- customer_credit DB migration (supabase/migrations/20260705000002_customer_credit.sql)
  + retail credit UI/API (pos-askbiz/app/retail/credit, customer-credit API)
- import-catalog API (app/api/pos/import-catalog)
- Supabase client/server updates (lib/supabase, pos-askbiz/lib/supabase)
- Purchase Orders tab, middleware, and i18n catalog/locale additions

Not independently tested here — review before relying on the migration.

+++ app/api/pos/inventory/route.ts
@@ -130,7 +130,7 @@ export async function PUT(req: NextRequest) {
     name:                (i.name || '').trim(),
     sale_price:          Number(i.sale_price) || 0,
     cost_price:          0,
-    stock_qty:           Math.max(0, parseInt(String(i.stock_qty)) || 0),
+    stock_qty:           Math.max(0, parseFloat(String(i.stock_qty)) || 0),  // numeric(10,3): keep decimals for weight items
     low_stock_threshold: 5,
     unit:                i.unit || 'item',
   })).filter(r => r.name)
